Fix Go mod/work config
Removes the go.work files and extraneous go.{mod,sum} files.
Additionally renames the virtual_fido package to vfido for easier use,
and removes /main.go, moving the demo app to /cmd/vfido-demo

Also cleans up the README.md for easier reading.
