+++ cmd/vfido-demo/demo.go
@@ -1,4 +1,4 @@
-package demo
+package main
 
 import (
 	"crypto/ecdsa"
@@ -13,14 +13,23 @@ import (
 	"os"
 	"strings"
 	"time"
-	"virtual_fido"
 
+	"github.com/bulwarkid/virtual-fido/vfido"
 	"github.com/spf13/cobra"
 )
 
-var vaultFilename string
-var vaultPassphrase string
-var identityID string
+func main() {
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Fprintf(os.Stderr, "error: %v\n", err)
+		os.Exit(1)
+	}
+}
+
+var (
+	vaultFilename   string
+	vaultPassphrase string
+	identityID      string
+)
 
 func checkErr(err error, message string) {
 	if err != nil {
@@ -40,7 +49,7 @@ func listIdentities(cmd *cobra.Command, args []string) {
 func deleteIdentity(cmd *cobra.Command, args []string) {
 	client := createClient()
 	identities := client.Identities()
-	targetIDs := make([]*virtual_fido.CredentialSource, 0)
+	targetIDs := make([]*vfido.CredentialSource, 0)
 	for _, id := range identities {
 		hexString := hex.EncodeToString(id.ID)
 		if strings.HasPrefix(hexString, identityID) {
@@ -69,7 +78,7 @@ func start(cmd *cobra.Command, args []string) {
 	runServer(client)
 }
 
-func createClient() virtual_fido.Client {
+func createClient() vfido.Client {
 	// ALL OF THIS IS INSECURE, FOR TESTING PURPOSES ONLY
 	authority := &x509.Certificate{
 		SerialNumber: big.NewInt(0),
@@ -90,9 +99,9 @@ func createClient() virtual_fido.Client {
 	checkErr(err, "Could not generate attestation CA cert bytes")
 	encryptionKey := sha256.Sum256([]byte("test"))
 
-	virtual_fido.SetLogOutput(os.Stdout)
+	vfido.SetLogOutput(os.Stdout)
 	support := ClientSupport{vaultFilename: vaultFilename, vaultPassphrase: vaultPassphrase}
-	return virtual_fido.NewClient(authorityCertBytes, privateKey, encryptionKey, &support, &support)
+	return vfido.NewClient(authorityCertBytes, privateKey, encryptionKey, &support, &support)
 }
 
 func printUsage(message string) {
@@ -137,11 +146,4 @@ func init() {
 	delete.Flags().StringVarP(&identityID, "identity", "", "", "Identity hash to delete")
 	delete.MarkFlagRequired("identity")
 	rootCmd.AddCommand(delete)
-}
-
-func Execute() {
-	if err := rootCmd.Execute(); err != nil {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(1)
-	}
 }
