+++ go.mod
@@ -1,0 +1,15 @@
+module github.com/bulwarkid/virtual-fido
+
+go 1.19
+
+require (
+	github.com/fxamacker/cbor/v2 v2.4.0
+	github.com/spf13/cobra v1.5.0
+	golang.org/x/crypto v0.0.0-20220829220503-c86fa9a7ed90
+)
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/x448/float16 v0.8.4 // indirect
+)
