--- go.work
@@ -1,6 +1,0 @@
-go 1.18
-
-use (
-	./demo
-	./virtual_fido
-)
