--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "demo"
-
-func main() {
-	demo.Execute()
-}
