+++ vfido/usb.go
@@ -1,4 +1,4 @@
-package virtual_fido
+package vfido
 
 import "fmt"
