--- virtual_fido/go.mod
@@ -1,10 +1,0 @@
-module virtual_fido
-
-go 1.18
-
-require (
-	github.com/fxamacker/cbor/v2 v2.4.0
-	golang.org/x/crypto v0.0.0-20220829220503-c86fa9a7ed90
-)
-
-require github.com/x448/float16 v0.8.4 // indirect
